pyproject.toml
[project]
name = "learningfocused"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "assemblyai>=0.46.0",
    "feedparser>=6.0.12",
    "ipykernel>=7.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
